// src/contoprix/generated.ts
export interface MenuItemComponentComponentSettings {
  label: string;
  url?: string | null;
  icon?: string | null;
  children?: MenuItemComponentComponentSettings[] | null;
}

export interface NavigationMenuComponentSettings {
  name?: string | null;
  items?: MenuItemComponentComponentSettings[] | null;
}